refactor: UpdateLengthRange에 정수로 변환된 Kilometer 값을 넘겨주도록 변경

+++ android/app/src/main/java/io/coursepick/coursepick/presentation/filter/CourseFilterBottomSheet.kt
@@ -103,8 +103,8 @@ fun CourseFilterBottomSheet(
                     onValueChange = { range: ClosedFloatingPointRange<Float> ->
                         onFilterAction(
                             CourseFilterAction.UpdateLengthRange(
-                                Kilometer(range.start.toDouble()),
-                                Kilometer(range.endInclusive.toDouble()),
+                                Kilometer(range.start.toInt()),
+                                Kilometer(range.endInclusive.toInt()),
                             ),
                         )
                     },
